Update
Update: controllers, view

+++ controllers/AdviceController.php
@@ -5,6 +5,8 @@ namespace app\controllers;
 use app\models\Advices;
 use app\models\AdvicesSearch;
 use app\models\AdvicesStudentsSearch;
+use app\models\Users;
+use Yii;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
@@ -73,6 +75,10 @@ class AdviceController extends Controller
      */
     public function actionCreate()
     {
+        if (Yii::$app->user->isGuest || Users::findOne(Yii::$app->user->id)->getTitleRoles() !== 'Admin') {
+            return $this->goHome();
+        }
+
         $model = new Advices();
 
         if ($this->request->isPost) {
@@ -99,6 +105,10 @@ class AdviceController extends Controller
      */
     public function actionUpdate($id)
     {
+        if (Yii::$app->user->isGuest || Users::findOne(Yii::$app->user->id)->getTitleRoles() !== 'Admin') {
+            return $this->goHome();
+        }
+
         $model = $this->findModel($id);
 
         if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
@@ -119,6 +129,10 @@ class AdviceController extends Controller
      */
     public function actionDelete($id)
     {
+        if (Yii::$app->user->isGuest || Users::findOne(Yii::$app->user->id)->getTitleRoles() !== 'Admin') {
+            return $this->goHome();
+        }
+        
         $this->findModel($id)->delete();
 
         return $this->redirect(['/']);
